Restructuring ETL Lite for Copilot
Copilot injects database credentials and other secrets into containers
via environment variables. To make this work with the ETL Lite binary, I
created a Python wrapper script for the etl lite binary to serve at the
entrypoint for the image. It grabs credentials from the environment,
writes them out to the settings.toml config file etl-lite is expecting,
then runs etl lite.

Copilot also uses FQDNs for services, with the domain part of the FQDN
provided via an environment variable. Wrapper script also grabs that and
sticks it in the config file.